+++ src/network_manager.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void networkSetup();
+void networkUpdate();